# tournois-sport-api/public/swagger.yaml
openapi: 3.0.3
info:
  title: Tournois Sport API
  description: API de gestion de tournois sportifs avec authentification, administration, joueurs, inscriptions, matchs et tournois.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Serveur local Symfony
...
paths:
...
  /api/tournaments:
    get:
      tags: [Tournois]
      summary: Lister les tournois
      responses:
        '200':
          description: Liste des tournois
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament'
    post:
      tags: [Tournois]
      summary: Créer un tournoi
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentCreateRequest'
      responses:
        '201':
          description: Tournoi créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          description: Données invalides
        '401':
          description: Non authentifié
  /api/tournaments/{id}:
    get:
      tags: [Tournois]
      summary: Afficher un tournoi
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Tournoi trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '404':
          description: Tournoi introuvable
    put:
      tags: [Tournois]
      summary: Modifier un tournoi
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentUpdateRequest'
      responses:
        '200':
          description: Tournoi modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé
        '404':
          description: Tournoi introuvable
    delete:
      tags: [Tournois]
      summary: Supprimer un tournoi
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Tournoi supprimé
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé
        '404':
          description: Tournoi introuvable
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
...
  schemas:
...
    TournamentCreateRequest:
      type: object
      required:
        - tournamentName
        - startDate
        - endDate
        - description
        - maxParticipants
        - sport
      properties:
        tournamentName:
          type: string
          example: Tournoi de Football
        startDate:
          type: string
          format: date
          example: '2026-06-01'
        endDate:
          type: string
          format: date
          example: '2026-06-10'
        location:
          type: string
          nullable: true
          example: Paris
        description:
          type: string
          example: Tournoi sportif organisé entre joueurs.
        maxParticipants:
          type: integer
          example: 16
        sport:
          type: string
          example: Football
        organizerId:
          type: integer
          example: 1
    TournamentUpdateRequest:
      type: object
      properties:
        tournamentName:
          type: string
          example: Tournoi de Football modifié
        startDate:
          type: string
          format: date
          example: '2026-06-01'
        endDate:
          type: string
          format: date
          example: '2026-06-10'
        location:
          type: string
          nullable: true
          example: Paris
        description:
          type: string
          example: Nouvelle description du tournoi.
        maxParticipants:
          type: integer
          example: 20
        sport:
          type: string
          example: Football
        winnerId:
          type: integer
          nullable: true
          example: 2
    Tournament:
      type: object
      properties:
        id:
          type: integer
          example: 1
        tournamentName:
          type: string
          example: Tournoi de Football
        startDate:
          type: string
          format: date
          example: '2026-06-01'
        endDate:
          type: string
          format: date
          example: '2026-06-10'
        location:
          type: string
          nullable: true
          example: Paris
        description:
          type: string
          example: Tournoi sportif organisé entre joueurs.
        maxParticipants:
          type: integer
          example: 16
        status:
          type: string
          example: a venir
        sport:
          type: string
          example: Football
        organizer:
          $ref: '#/components/schemas/UserShort'
        winner:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserShort'
...
    UserShort:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: chloe
        emailAddress:
          type: string
          format: email
          example: chloe@example.com
...